Solving bugs 8
