+++ app/src/main/java/com/github/sdp_begreen/begreen/models/TrashPhotoMetadata.kt
@@ -37,18 +37,4 @@ data class TrashPhotoMetadata(
             return arrayOfNulls(size)
         }
     }
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is TrashPhotoMetadata) return false
-
-        // TrashPhotoMetadata is uniquely defined by the picture id because if we compare all fields, some of them
-        // are obsolete in the database. Actually this was the case in our case.
-
-        return pictureId == other.pictureId
-    }
-
-    override fun hashCode(): Int {
-        return pictureId?.hashCode() ?: 0
-    }
 }
